chore: update version 2.9.4 -> 2.9.5 beta (#1185)
* fix: 修复热力图点被拉伸的问题 (#1169)

* feat: 去处事件节流

* fix: 修复热力图点被拉伸的问题

* style: lint style

* chore: update version 2.8.45 -> 2.8.46 (#1170)

* feat: 去处事件节流

* fix: 修复热力图点被拉伸的问题

* style: lint style

* chore: update version 2.8.45 -> 2.8.46

* chore: 透出组件类型定义 (#1156)

* feat: sprite 粒子增加运动边界参数 spriteBottom (#1174)

* feat: 去处事件节流

* feat: sprite 补充参数

* chore: 透出组件类型定义 (#1175)

* chore: 透出组件类型定义

* chore: 修复类型导出问题

* chore: 代码格式化

* chore: update version 2.8.46 -> 2.8.47 (#1176)

* feat: 去处事件节流

* chore: update version 2.8.46 -> 2.8.47

* chore: update version 2.9.4 -> 2.9.5
