+++ lerna.json
@@ -15,7 +15,7 @@
       "registry": "https://registry.npmjs.org"
     }
   },
-  "version": "2.9.4",
+  "version": "2.9.5",
   "npmClient": "yarn",
   "useWorkspaces": true,
   "publishConfig": {
