+++ packages/component/src/css/l7.css
@@ -238,7 +238,7 @@
 /* general toolbar styles */
 
 .l7-bar {
-  box-shadow: 0 1px 5px rgba(0, 0, 0, 0.65);
+  box-shadow: 0 0 1px rgba(0, 0, 0, 0.3);
   border-radius: 4px;
 }
 .l7-bar a,
