+++ packages/component/src/marker-layer.ts
@@ -13,24 +13,8 @@ import { merge } from 'lodash';
 // @ts-ignore
 // tslint:disable-next-line:no-submodule-imports
 import Supercluster from 'supercluster/dist/supercluster';
+import { IMarkerLayerOption, IMarkerStyleOption } from './interface';
 import Marker from './marker';
-type CallBack = (...args: any[]) => any;
-interface IMarkerStyleOption {
-  element?: CallBack;
-  style: { [key: string]: any } | CallBack;
-  className: string;
-  field?: string;
-  method?: 'sum' | 'max' | 'min' | 'mean';
-  radius: number;
-  maxZoom: number;
-  minZoom: number;
-  zoom: number;
-}
-
-interface IMarkerLayerOption {
-  cluster: boolean;
-  clusterOption: Partial<IMarkerStyleOption>;
-}
 
 interface IPointFeature {
   geometry: {
@@ -39,6 +23,7 @@ interface IPointFeature {
   };
   properties: any;
 }
+
 export default class MarkerLayer extends EventEmitter {
   private markers: IMarker[] = [];
   private markerLayerOption: IMarkerLayerOption;
