+++ packages/l7/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "A Large-scale WebGL-powered Geospatial Data Visualization",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -25,13 +25,13 @@
   "author": "antv",
   "license": "MIT",
   "dependencies": {
-    "@antv/l7-component": "2.9.4",
-    "@antv/l7-core": "2.9.4",
-    "@antv/l7-layers": "2.9.4",
-    "@antv/l7-maps": "2.9.4",
-    "@antv/l7-scene": "2.9.4",
-    "@antv/l7-source": "2.9.4",
-    "@antv/l7-utils": "2.9.4",
+    "@antv/l7-component": "2.9.5",
+    "@antv/l7-core": "2.9.5",
+    "@antv/l7-layers": "2.9.5",
+    "@antv/l7-maps": "2.9.5",
+    "@antv/l7-scene": "2.9.5",
+    "@antv/l7-source": "2.9.5",
+    "@antv/l7-utils": "2.9.5",
     "@babel/runtime": "^7.7.7"
   },
   "gitHead": "684ba4eb806a798713496d3fc0b4d1e17517dc31",
