+++ packages/layers/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-layers",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "L7's collection of built-in layers",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -25,9 +25,9 @@
   "author": "xiaoiver",
   "license": "ISC",
   "dependencies": {
-    "@antv/l7-core": "2.9.4",
-    "@antv/l7-source": "2.9.4",
-    "@antv/l7-utils": "2.9.4",
+    "@antv/l7-core": "2.9.5",
+    "@antv/l7-source": "2.9.5",
+    "@antv/l7-utils": "2.9.5",
     "@babel/runtime": "^7.7.7",
     "@mapbox/martini": "^0.2.0",
     "@turf/helpers": "^6.1.4",
