+++ packages/layers/src/core/interface.ts
@@ -197,6 +197,7 @@ export interface IGeometryLayerStyleOptions {
   spriteCount?: number;
   spriteSpeed?: number;
   spriteTop?: number;
+  spriteBottom?: number;
   spriteUpdate?: number;
   spriteScale?: number;
