+++ packages/map/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-map",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "l7 map",
   "keywords": [],
   "author": "thinkinggis <lzx199065@gmail.com>",
@@ -37,7 +37,7 @@
   },
   "homepage": "https://github.com/antvis/L7#readme",
   "dependencies": {
-    "@antv/l7-utils": "2.9.4",
+    "@antv/l7-utils": "2.9.5",
     "@babel/runtime": "^7.7.7",
     "@mapbox/point-geometry": "^0.1.0",
     "@mapbox/unitbezier": "^0.0.0",
