+++ packages/renderer/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-renderer",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -26,7 +26,7 @@
     "gl": "^4.4.0"
   },
   "dependencies": {
-    "@antv/l7-core": "2.9.4",
+    "@antv/l7-core": "2.9.5",
     "@babel/runtime": "^7.7.7",
     "inversify": "^5.0.1",
     "l7regl": "^0.0.20",
