+++ packages/scene/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-scene",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -23,12 +23,12 @@
   "author": "xiaoiver",
   "license": "ISC",
   "dependencies": {
-    "@antv/l7-component": "2.9.4",
-    "@antv/l7-core": "2.9.4",
-    "@antv/l7-layers": "2.9.4",
-    "@antv/l7-maps": "2.9.4",
-    "@antv/l7-renderer": "2.9.4",
-    "@antv/l7-utils": "2.9.4",
+    "@antv/l7-component": "2.9.5",
+    "@antv/l7-core": "2.9.5",
+    "@antv/l7-layers": "2.9.5",
+    "@antv/l7-maps": "2.9.5",
+    "@antv/l7-renderer": "2.9.5",
+    "@antv/l7-utils": "2.9.5",
     "@babel/runtime": "^7.7.7",
     "inversify": "^5.0.1",
     "mapbox-gl": "^1.2.1",
