+++ packages/source/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-source",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "",
   "main": "lib/index.js",
   "module": "es/index.js",
@@ -26,8 +26,8 @@
   "license": "ISC",
   "dependencies": {
     "@antv/async-hook": "^2.1.0",
-    "@antv/l7-core": "2.9.4",
-    "@antv/l7-utils": "2.9.4",
+    "@antv/l7-core": "2.9.5",
+    "@antv/l7-utils": "2.9.5",
     "@babel/runtime": "^7.7.7",
     "@mapbox/geojson-rewind": "^0.4.0",
     "@mapbox/vector-tile": "^1.3.1",
