+++ packages/utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/l7-utils",
-  "version": "2.9.4",
+  "version": "2.9.5",
   "description": "",
   "main": "lib/index.js",
   "module": "es/index.js",
